requirements.txt
pygame==2.5.2
pillow>=9.0.0
numpy
pyinstaller